+++ pkg/testing/mock-supplychain.go
@@ -174,3 +174,47 @@ func (p *BlueprintInputsFile) GetBlueprintInputs() (*Inputs, error) {
 
 	return &inputs, nil
 }
+
+func (i *TemplateTestGivens) createTemplatingContext(workload v1alpha1.Workload, params map[string]apiextensionsv1.JSON) (map[string]interface{}, error) {
+	var (
+		inputs *Inputs
+		err    error
+	)
+
+	inputs = &Inputs{}
+
+	if i.MockSupplyChain.BlueprintInputs != nil {
+		inputs, err = i.MockSupplyChain.BlueprintInputs.GetBlueprintInputs()
+		if err != nil {
+			return nil, fmt.Errorf("get supply chain inputs: %w", err)
+		}
+	}
+
+	templatingContext := map[string]interface{}{
+		"workload": workload,
+		"params":   params,
+		"sources":  inputs.Sources,
+		"images":   inputs.Images,
+		"configs":  inputs.Configs,
+		//"deployment": // not implemented yet,
+	}
+
+	if len(inputs.Sources) == 1 {
+		for _, source := range inputs.Sources {
+			templatingContext["source"] = &source
+		}
+	}
+
+	if len(inputs.Images) == 1 {
+		for _, image := range inputs.Images {
+			templatingContext["image"] = image.Image
+		}
+	}
+
+	if len(inputs.Configs) == 1 {
+		for _, config := range inputs.Configs {
+			templatingContext["config"] = config.Config
+		}
+	}
+	return templatingContext, nil
+}
